tr: hidden is an object property
Hidden is now managed at object level, not at task level. This spares
the need to find the raw task twice (once for the hidden attribute and
once for the full task descriptor).

Note that while objects may be hidden, there ability to be so is
enforced by the schema. As of now, only tasks can be hidden and
containers (the only other current object), can not.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>