tr: seperate task and container objects
Task descriptors no long includes the expanded container information
internally. Instead it just keeps a reference to the container name
without any expansion.

Clients of the Task object need to explicitly request a container
descriptor in order to get the container information. As of now, this is
only relevant to the Task class, but might be useful in the future. Plus
it make the code flow simpler, readable and easier to maintain.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>