tr: run container as root (sudo) setting
Add the ability to run container based commands as root by prefixing
the container run/exec command with sudo. On a well configured system,
role and task (i.e. 'real' task), this shouldn't be used, but nothing is
really that simple, and this isn't the taskrunner tool responsibility to
begin with, so add it; can be handy.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>