tr: add meta data for object
Add a "meta" sub-object to the task dictionary. Allows adding some data
to the task that task-runner infrastructure is agnostic to. Mainly
designed to have some runtime test support (e.g. Determine which tasks
should be allowed to fail.) In the future it, can be used to whatever
the user wants to, although right now it isn't very helpful as this data
isn't available in runtime, but might expose it in the future through a
environment variable.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>