tr: config file include re-haul (scehma v3.0)
Completely rework the way configuration files works in terms of
inclusion. Before, the configuration file was read (if found), and the
global file was also read (if found).

With this patch:
 - Local configuration is searched as before
 - If found, no global file is read
 - If not found, a global configuration file is searched
 - Each configuration can have a include section, where it which other
   files are to be included for this file. This files are read one by
   one, each one overrides the previous. The original files overrides
   all the other configuration settings. To mimic the old behavior, have
   a local file include the global file explicitly.
 - Included files may have include directives themselves.

This allows a much more flexible inheritance-like behavior, where a
given tasks configuration might inherit other configurations that are
relevant for it, and override what's relevant. This also greatly
simplify the testing framework.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>