test/050: split to Ubuntu and Rocky
Have the old 050 test run in two variants, one for Ubuntu 20.04 (050)
and one for Rocky 8.5 (051). Both inherit from the same abstract base
task. In addition to the simple environment variable print out of the
old test, also print the os-release file , make sure the test indeed run
in a container.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>