tr: return 255 on internal error
Use 255 return code for internal task error. So it differentiated from
errors returned from the tasks themselves. Those could still return 255,
but what can you do.

This allows the tests suite to print the output file in case of internal
errors instead of the log file, which was confusing as hell when
debugging.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>