test: expose task binary through environment
As some of the test tasks actually run the task executable, some hoop
jumping was used to get the right task binary to be used when running
tests, as the test suite might be run for a 'task' in the repository, or
one provided by the package.

Instead, let the test suite expose the binary path through an
environment variable, and use 'task' own capability to translate it into
a variable to get the right executable path.

Also provides an additional coverage for the entire TR and environment
variables referencing and expansion.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>