tr: inherit environment by default (v0.6.0, s4.4)
Modifies and streamlines the way environment variables are used.

Previously if no variables were defined for a task, the current set of
environment variables were passed to the task commands, but if any such
variables *were* defined, a dictionary with only those variables was
passed, ignoring the current environment variables set completely, which
was inconsistent.

This patches changes so by default, the existing variables are always
passed to the task. If task specific environment variables are set
(either in the task descriptor or the from CLI), they are appended (or
override), to the current set.

In addition, a new boolean schema setting named "env_inherit" is added,
defaulted to true. If set to false, current environment variables are
ignored, and only task specific variables are passed to the task.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>