tr: kill schema version
Stop with the schema version. It only causes grief. Validation is always
against what the current version knows. If something is wrong, then it
need to be fixed, and that's it.

Forces the schema files naming to change as well - instead of the
previous schema version, just use the last TR version that modified the
schema. Users are expected to validate their schemas according to the
version they use (if they even choose to do so).

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>