test: don't treat 255 as special rc
Use 255 return code just like any other code that can be allowed. We can
never know if the wrong RC is a result of an internal TR bug, or a
command that exited with 255, so on wrong RC returned by TR, print out
both the output file if there was such, and the log file.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>