tr: revised task inheritance
Previous task inheritance mechanism didn't handle environment, container
environment and container volumes correctly. These set of settings
(along the already handled variables settings), require special
handling, as a derived task might want to override or extend the base
task it inherits for. With this patch, the default inheritance behavior
is to always extend the setting in question, while providing a boolean
"inherit_<SETTING>" (e.g. "inherit_env"), that allows the task author to
ignore this specific base setting.

All the these handling are done in a generic way in the descriptor
generation. "inherit_<SETTING>" are removed from the returned
descriptor.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>