tr: migrate to pydantic (major change)
Move the entire scheme validation to by pydantic based rather than the
old explicit schema based.

This actually modifies the entire code base in a rather drastic way, and
the code is no longer compatible with previous versions.  The motivation
for this commit was pydantic, it is, in fact, a rather extensive rewrite
of the code.

- Pre an post commands are removed. There is now only commands.
- No longer 'meta' entries in the tasks. These were born for testings,
  which is about to migrated to Xeet.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>